New generator: latex logic generator.
This generates just the latex without worrying about line breaking, hoping that Latex would. It doesn't, on it's. There
is parameterization to allow for use of two packages(breqn and autobreak) in the form of two instance
variables. formula-format is a format string, and format, when generating, will use this passing two arguments - the
name of the axiom, and the latex rendering of it. insert-line-breaks supports the autobreak package, which uses a line
break as a hint on where to start a new line. Setting insert-line-breaks to t will result in a few places (determined by
trial and error) that usually result in the formulas printing without extending off the right of the page.

proof.lisp
Added options to run-proof :print-axiom-names :print-formulas :print-executed-form, all default nil, which when t do
what they say. Fixed a bug in non-entailment proof where :timeout wasn't being returned when appropriate.
Pretty: When both names and formula are written, and we're in slide, the names are in blue for better
readability.

logic.asd: add latex-logic-generator