Checking a model using clausetester finally works. The call is (check-model model-spec theory-spec) The function will return 2 values, first is either :SATISFYING-MODEL or :FAILED.  The second is a list of the formulas (suitable for pprint-spec-axioms) of either the succeeding or failing formulas respectively.
To do this we construct an interpretation in the form that mace4 does.
The program clausetester (https://www.cs.unm.edu/~mccune/prover9/manual/2009-11A/) takes that interpretation and a bunch of formulas
and prints a line for each formula along with whether the interpretation was consistent with it ("1") or not <empty>

A complication is that we've used the set_prolog_variables convention for prover9 and mace4 which makes names starting
with a capital letter be consider free variables instead of constants. Unfortunately clausetester doesn't have a way to
indicate that, so we have to mangle all the constant (and for the ride the predicate) names by prefixing them with an
"_". Nothing that should be visible, but a frustrating few hours until I figured this out.

For now *prover9-uparrow-initialcap* is set to nil. I'll try to fix that.