dl-query.lisp bug in property-descendents. wasn't doing recursively.
domain-and-range.lisp
New function
(property-applicable-to-domain-range prop domain range &key (kb *default-kb*) (ignore-owl-thing t) (explicitish t))

Check if a property can relate instances of domain and range if explicitish is t then accept if one of the named
subclasses of the domain/range is a subclass of the target domain/range if explicitish is nil then accept if
intersection of target domain/range and prop domain/range is satisfiable if ignore-owl-thing is t then domains/ranges
that are owl:Thing are ignored under the assumption that the relation domain/range wasn't explicitly stated

owlapi.lisp
Make property query handle unnamed inverse properties.

release.lisp
I think it is broken :-(
Minor bugfixes.